+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support translation of optional types (`Optional<T>`, `OptionalInt`, `OptionalLong` and `OptionalDouble`).
   ([#15](https://github.com/EvidentSolutions/apina/issues/15))
 - Support configuring base URL for API calls. ([#13](https://github.com/EvidentSolutions/apina/issues/13))
+- Use black-box translation for classes with `@JsonValue` ([#11](https://github.com/EvidentSolutions/apina/issues/11))
 
 ## Breaking changes