+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## x.y.z (yyyy-mm-dd)
+## 0.3.0 (2015-08-19)
 
 ### New features